Крупное обновление
Настройки. Game Over. Изменения интерфейса. Апгрейд статов игрока. Сохранение/загрузка прогресса

+++ Assets/Main/Scripts/EnemyData.cs
@@ -1,6 +1,3 @@
-using System.Collections;
-using System.IO;
-using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "EnemyData", menuName = "Scriptable Objects/Enemy Data")]
@@ -8,7 +5,7 @@ public class EnemyData : ScriptableObject {
     public string Name;
     public GameObject Prefab;
     public int CostOfCoins;
-    public float Health = 100;
-    public float Damage = 1;
+    public int Health = 100;
+    public int Damage = 1;
     public float AttackSpeed = 1;
 }
